% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_field.R
\name{check_field}
\alias{check_field}
\alias{check_integer}
\alias{check_number}
\alias{check_string}
\alias{check_boolean}
\alias{check_date}
\title{Check if a vector is valid given a field descriptor}
\usage{
check_field(
  x,
  fielddescriptor,
  constraints = TRUE,
  tolerance = sqrt(.Machine$double.eps)
)

check_integer(
  x,
  fielddescriptor,
  constraints = TRUE,
  tolerance = sqrt(.Machine$double.eps)
)

check_number(x, fielddescriptor, constraints = TRUE)

check_string(x, fielddescriptor, constraints = TRUE)

check_boolean(x, fielddescriptor, constraints = TRUE)

check_date(x, fielddescriptor, constraints = TRUE)
}
\arguments{
\item{x}{vector to test}

\item{fielddescriptor}{field descriptor to test the vector against}

\item{constraints}{also check relevant constraints in the field descriptor.}

\item{tolerance}{numerical tolerance used in some of the tests}
}
\value{
Returns \code{TRUE} when the field is valid. Returns a character vector with
length >= 1 if the field is not valid. The text in the character values 
indicates why the field is not valid.
}
\description{
Check if a vector is valid given a field descriptor
}
\seealso{
Use \code{\link{isTRUE}} to check if the test was successful.
}
