% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_boolean.R
\name{complete_fielddescriptor_boolean}
\alias{complete_fielddescriptor_boolean}
\title{Add required fields to the field descriptor for an boolean column}
\usage{
complete_fielddescriptor_boolean(fielddescriptor)
}
\arguments{
\item{fielddescriptor}{should be a list.}
}
\value{
Returns \code{fielddescriptor} with the required fields added.
}
\description{
Add required fields to the field descriptor for an boolean column
}
