% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_date.R
\name{complete_fielddescriptor_date}
\alias{complete_fielddescriptor_date}
\title{Add required fields to the fielddescriptor for a date column}
\usage{
complete_fielddescriptor_date(fielddescriptor)
}
\arguments{
\item{fielddescriptor}{should be a list.}
}
\value{
Returns \code{fielddescriptor} with the required fields added.
}
\description{
Add required fields to the fielddescriptor for a date column
}
