% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_datetime.R
\name{complete_fielddescriptor_datetime}
\alias{complete_fielddescriptor_datetime}
\title{Add required fields to the fielddescriptor for a datetime column}
\usage{
complete_fielddescriptor_datetime(fielddescriptor)
}
\arguments{
\item{fielddescriptor}{should be a list.}
}
\value{
Returns \code{fielddescriptor} with the required fields added.
}
\description{
Add required fields to the fielddescriptor for a datetime column
}
