% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_integer.R
\name{complete_fielddescriptor_integer}
\alias{complete_fielddescriptor_integer}
\title{Add required fields to the field descriptor for an integer column}
\usage{
complete_fielddescriptor_integer(fielddescriptor)
}
\arguments{
\item{fielddescriptor}{should be a list.}
}
\value{
Returns \code{fielddescriptor} with the required fields added.
}
\description{
Add required fields to the field descriptor for an integer column
}
