% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_number.R
\name{complete_fielddescriptor_number}
\alias{complete_fielddescriptor_number}
\title{Add required fields to the field descriptor for an number column}
\usage{
complete_fielddescriptor_number(fielddescriptor)
}
\arguments{
\item{fielddescriptor}{should be a list.}
}
\value{
Returns \code{fielddescriptor} with the required fields added.
}
\description{
Add required fields to the field descriptor for an number column
}
