% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_string.R
\name{complete_fielddescriptor_string}
\alias{complete_fielddescriptor_string}
\title{Add required fields to the field descriptor for a string column}
\usage{
complete_fielddescriptor_string(fielddescriptor)
}
\arguments{
\item{fielddescriptor}{should be a list.}
}
\value{
Returns \code{fielddescriptor} with the required fields added.
}
\description{
Add required fields to the field descriptor for a string column
}
