% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_yearmonth.R
\name{complete_fielddescriptor_yearmonth}
\alias{complete_fielddescriptor_yearmonth}
\title{Add required fields to the fielddescriptor for a yearmonth column}
\usage{
complete_fielddescriptor_yearmonth(fielddescriptor)
}
\arguments{
\item{fielddescriptor}{should be a list.}
}
\value{
Returns \code{fielddescriptor} with the required fields added.
}
\description{
Add required fields to the fielddescriptor for a yearmonth column
}
