% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_boolean.R
\name{complete_schema_boolean}
\alias{complete_schema_boolean}
\title{Add required fields to the schema for an boolean column}
\usage{
complete_schema_boolean(schema)
}
\arguments{
\item{schema}{should be a list.}
}
\value{
Returns \code{schema} with the required fields added.
}
\description{
Add required fields to the schema for an boolean column
}
