% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_date.R
\name{complete_schema_date}
\alias{complete_schema_date}
\title{Add required fields to the schema for a date column}
\usage{
complete_schema_date(schema)
}
\arguments{
\item{schema}{should be a list.}
}
\value{
Returns \code{schema} with the required fields added.
}
\description{
Add required fields to the schema for a date column
}
