% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_integer.R
\name{complete_schema_integer}
\alias{complete_schema_integer}
\title{Add required fields to the schema for an integer column}
\usage{
complete_schema_integer(schema)
}
\arguments{
\item{schema}{should be a list.}
}
\value{
Returns \code{schema} with the required fields added.
}
\description{
Add required fields to the schema for an integer column
}
