% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_number.R
\name{complete_schema_number}
\alias{complete_schema_number}
\title{Add required fields to the schema for an number column}
\usage{
complete_schema_number(schema)
}
\arguments{
\item{schema}{should be a list.}
}
\value{
Returns \code{schema} with the required fields added.
}
\description{
Add required fields to the schema for an number column
}
