% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_string.R
\name{complete_schema_string}
\alias{complete_schema_string}
\title{Add required fields to the schema for an string column}
\usage{
complete_schema_string(schema)
}
\arguments{
\item{schema}{should be a list.}
}
\value{
Returns \code{schema} with the required fields added.
}
\description{
Add required fields to the schema for an string column
}
