% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{newcontributor}
\alias{newcontributor}
\alias{dpaddcontributor}
\alias{dpaddcontributor<-}
\title{Creating and Adding Contributors to a Data Package}
\usage{
newcontributor(
  title,
  role = c("contributor", "author", "publisher", "maintainer", "wrangler"),
  path = NULL,
  email = NULL,
  organisation = NULL
)

dpaddcontributor(x, contributor)

dpaddcontributor(x) <- value
}
\arguments{
\item{title}{A length 1 character vector with the full nam of the
contributor.}

\item{role}{The role of the contributor}

\item{path}{A URL to e.g. a home page of the contributor}

\item{email}{The email address of the contributor}

\item{organisation}{The orgination the contributor belongs to.}

\item{x}{The Data Package to which the contributor has to be added.}

\item{contributor}{a contributor object}

\item{value}{a contributor object}
}
\value{
\code{newcontributor} returns a list with the given properties. This function
is meant to assist in creating valid contributors.
}
\description{
Creating and Adding Contributors to a Data Package
}
\examples{
dp <- opendatapackage(system.file(package = "datapackage", "examples/iris")) 
dpcontributors(dp)
dpcontributors(dp) <- list(
  newcontributor("John Doe", email = "j.doe@somewhere.org"),
  list(title = "Jane Doe", role = "maintainer")
)
dpaddcontributor(dp) <- newcontributor("Janet Doe")

}
