% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_write.R
\name{csv_write}
\alias{csv_write}
\title{Write data of data resource to CSV-file}
\usage{
csv_write(x, resourcename, datapackage, use_fwrite = FALSE, ...)
}
\arguments{
\item{x}{\code{data.frame} with the data to write}

\item{resourcename}{name of the data resource in the data package.}

\item{datapackage}{the Data Package to which the file should be written.}

\item{use_fwrite}{write the file using \code{fwrite} from the
\code{data.table} package.}

\item{...}{ignored for now}
}
\value{
The function doesn't return anything. It is called for it's side effect of
creating CSV-files in the directory of the data package.
}
\description{
Write data of data resource to CSV-file
}
