% Generated by roxygen2 (4.0.2): do not edit by hand
\name{derive_schema}
\alias{derive_schema}
\title{Derive a schema from a table}
\usage{
derive_schema(tbl, ...)
}
\arguments{
\item{tbl}{table, can be a \code{data.frame} or a \code{\link{tbl}} object.}
}
\description{
Derive a schema from a table
}

