% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_field.R
\name{dp_field}
\alias{dp_field}
\title{Get the Field Descriptor associated with a certain field in a Data Resource}
\usage{
dp_field(x, field_name)
}
\arguments{
\item{x}{a \code{dataresource} or \code{tableschema} object.}

\item{field_name}{length one character vector with the name of the field.}
}
\value{
An object of type \code{fielddescriptor}.
}
\description{
Get the Field Descriptor associated with a certain field in a Data Resource
}
