% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_fieldnames.R
\name{dp_field_names}
\alias{dp_field_names}
\title{List the fields in a Data Resource}
\usage{
dp_field_names(x)
}
\arguments{
\item{x}{object for which to get the field names. This can either be a Data
Resource or Table Schema.}
}
\value{
Returns a character vector with the fields in the Data Resource.
}
\description{
List the fields in a Data Resource
}
