% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_get_connection.R
\name{dp_get_connection}
\alias{dp_get_connection}
\title{Get a connection to the data belonging to a Data Resource}
\usage{
dp_get_connection(x, ...)
}
\arguments{
\item{x}{Can either be a Data Resource or Data Package.}

\item{...}{Extra arguments are passed on to \code{\link{dp_get_data}}.}
}
\value{
Depending on the type of Data Resource a connection to the data is returned.
Some readers will return the data set as a \code{data.frame}.
}
\description{
Get a connection to the data belonging to a Data Resource
}
\details{
When \code{x} is a Data Package a additional argument \code{resource_name} is
needed to identify the correct Data Resource. See \code{\link{dp_get_data}}. 

This function calls \code{dp_get_data} with an additional 
\code{as_connection = TRUE)} argument.
}
