% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_get_datapackage.R
\name{dp_get_datapackage}
\alias{dp_get_datapackage}
\title{Get the Data Package associated with the object}
\usage{
dp_get_datapackage(x)
}
\arguments{
\item{x}{the object for which to determine the associated Data Package}
}
\value{
Returns a Data Resource object, or returns \code{NULL} when none could be
found.
}
\description{
Get the Data Package associated with the object
}
\details{
This method can, of course, only determine the Data Package when this
information is stored in one of the attributes of the object. This can be
either be a \code{datapackage} attribute or an \code{dataresource}
attribute.
}
