% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_load_from_datapackage.R
\name{dp_load_from_datapackage}
\alias{dp_load_from_datapackage}
\title{Quickly read a dataset from a Data Package}
\usage{
dp_load_from_datapackage(path, resource_name, ...)
}
\arguments{
\item{path}{the directory with the Data Package}

\item{resource_name}{the name of the Data Resource. When omitted the Data Resource
with the same name as the Data Package is read in and when no such resource
exists the first Data Resource is read in.}

\item{...}{passed on to \code{\link{dp_get_data}}.}
}
\value{
Returns a dataset. Usually a \code{data.frame}.
}
\description{
Quickly read a dataset from a Data Package
}
\details{
This function is a wrapper around \code{\link{open_datapackage}} and
\code{\link{dp_get_data}}. It offers a quick way to read in a dataset from a
Data Package.
}
