% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_properties.R
\name{dp_properties}
\alias{dp_properties}
\alias{dp_properties.readonlydatapackage}
\alias{dp_properties.editabledatapackage}
\alias{dp_properties.dataresource}
\alias{dp_properties.tableschema}
\title{Get a list of properties defined for the object}
\usage{
dp_properties(x)

\method{dp_properties}{readonlydatapackage}(x)

\method{dp_properties}{editabledatapackage}(x)

\method{dp_properties}{dataresource}(x)

\method{dp_properties}{tableschema}(x)
}
\arguments{
\item{x}{the object for which to obtain the properties}
}
\value{
Returns a character vector (possibly zero length) with the names of the
properties.
}
\description{
Get a list of properties defined for the object
}
\seealso{
The \code{\link{dp_property}} method can be used to get the values of the
properties.
}
