% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_property.R
\name{dp_property}
\alias{dp_property}
\alias{dp_property.readonlydatapackage}
\alias{dp_property.editabledatapackage}
\alias{dp_property<-}
\alias{dp_property<-.readonlydatapackage}
\alias{dp_property<-.editabledatapackage}
\alias{dp_property.dataresource}
\alias{dp_property<-.dataresource}
\alias{dp_property.tableschema}
\alias{dp_property<-.tableschema}
\alias{dp_property.fielddescriptor}
\alias{dp_property<-.fielddescriptor}
\title{Get and set properties of Data Packages and Data Resources}
\usage{
dp_property(x, attribute)

\method{dp_property}{readonlydatapackage}(x, attribute)

\method{dp_property}{editabledatapackage}(x, attribute)

dp_property(x, attribute) <- value

\method{dp_property}{readonlydatapackage}(x, attribute) <- value

\method{dp_property}{editabledatapackage}(x, attribute) <- value

\method{dp_property}{dataresource}(x, attribute)

\method{dp_property}{dataresource}(x, attribute) <- value

\method{dp_property}{tableschema}(x, attribute)

\method{dp_property}{tableschema}(x, attribute) <- value

\method{dp_property}{fielddescriptor}(x, attribute)

\method{dp_property}{fielddescriptor}(x, attribute) <- value
}
\arguments{
\item{x}{a \code{datapackage} or \code{dataresource} object.}

\item{attribute}{a length 1 character vector with the name of the property.}

\item{value}{the new value of the property.}
}
\value{
Either returns the property or modifies the object.
}
\description{
Get and set properties of Data Packages and Data Resources
}
\seealso{
See \code{\link{dp_name}} etc. for methods for specific properties for Data
Packages and \code{\link{dp_encoding}} etc. for specific properties for Data
Resources. These specific methods also check if the input is valid for the
given property.
}
