% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_resource_names.R
\name{dp_resource_names}
\alias{dp_resource_names}
\title{Get the names of the resources in a Data Package}
\usage{
dp_resource_names(dp)
}
\arguments{
\item{dp}{A \code{datapackage} object.}
}
\value{
Returns a character vector with the names of the data resources in the Data
Package.
}
\description{
Get the names of the resources in a Data Package
}
