% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_resources.R
\name{dp_resources<-}
\alias{dp_resources<-}
\title{Modify a set of Data Resources in a Data Package}
\usage{
dp_resources(x) <- value
}
\arguments{
\item{x}{a \code{datapackage} object}

\item{value}{a \code{dataresource} object or a list of \code{dataresource}
objects .}
}
\value{
Returns a modified \code{x}.
}
\description{
Modify a set of Data Resources in a Data Package
}
