% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_save_as_datapackage.R
\name{dp_save_as_datapackage}
\alias{dp_save_as_datapackage}
\title{Save a dataset as a Data Package}
\usage{
dp_save_as_datapackage(
  data,
  path,
  name,
  categories_type = c("regular", "resource")
)
}
\arguments{
\item{data}{the data.frame with the data to save}

\item{path}{directory in which to create the datapackage}

\item{name}{name of the Data Resource. When omitted a name is generated.}

\item{categories_type}{how should categories be stored. See 
\code{\link{dp_generate_fielddescriptor}}.}
}
\value{
Does not return anything. Called for the side effect of creating a directory
and creating a number of files in this directory. Together these form a
complete Data Package.
}
\description{
Save a dataset as a Data Package
}
\details{
This function is a wrapper function around \code{\link{new_datapackage}},
\code{\link{dp_generate_dataresource}} and \code{\link{dp_write_data}}. These
functions are called with the default arguments. This allows for a quick way
to save a data set with any necessary data needed to read the dataset.
}
