% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_boolean.R
\name{dp_to_boolean}
\alias{dp_to_boolean}
\title{Convert a vector to 'boolean' using the specified field descriptor}
\usage{
dp_to_boolean(x, fielddescriptor = list(), ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{fielddescriptor}{the field descriptor for the field.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{logical} vector with \code{fielddescriptor} added as
the 'fielddescriptor' attribute.
}
\description{
Convert a vector to 'boolean' using the specified field descriptor
}
\details{
When \code{fielddescriptor} is missing a default field descriptor is
generated using \code{\link{complete_fielddescriptor_boolean}}.
}
