% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_date.R
\name{dp_to_date}
\alias{dp_to_date}
\title{Convert a vector to 'date' using the specified field descriptor}
\usage{
dp_to_date(x, fielddescriptor = list(), ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{fielddescriptor}{the field descriptor for the field.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{Date} vector with \code{fielddescriptor} added as the
'fielddescriptor' attribute.
}
\description{
Convert a vector to 'date' using the specified field descriptor
}
\details{
When \code{fielddescriptor} is missing a default field descriptor is
generated.
}
