% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpapplyschema.R
\name{dpapplyschema}
\alias{dpapplyschema}
\title{Convert columns of data.frame to their correct types using table schema}
\usage{
dpapplyschema(dta, resource, to_factor = FALSE, ...)
}
\arguments{
\item{dta}{a \code{data.frame} or \code{data.table}.}

\item{resource}{an object with the Data Resource of the data set.}

\item{to_factor}{convert columns to factor if the field descriptor a
\code{codelist} field for the column.}

\item{...}{additional arguments are passed on to the \code{to_<fieldtype>} 
functions (e.g. \code{\link{to_number}}).}
}
\description{
Convert columns of data.frame to their correct types using table schema
}
\details{
Converts each column in \code{dta} to the correct R-type using the type
information in the table schema. For example, if the original column type in
\code{dta} is a character vector and the table schema specifies that the field is
of type number, the column is converted to numeric using the decimal
separator and thousands separator specified in the field descriptor (or default values
for these if not).
}
\seealso{
This function calls conversion functions for each of the columns, see 
\code{\link{to_number}}, \code{\link{to_boolean}}, \code{\link{to_integer}}, 
and \code{\link{to_date}}.
}
