% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcategorieslist.R
\name{dpcategorieslist}
\alias{dpcategorieslist}
\alias{dpcategorieslist.default}
\alias{dpcategorieslist.fielddescriptor}
\title{Get the a data.frame with the categories for a variable.}
\usage{
dpcategorieslist(x, ...)

\method{dpcategorieslist}{default}(
  x,
  fielddescriptor = attr(x, "fielddescriptor"),
  datapackage = dpgetdatapackage(fielddescriptor),
  ...
)

\method{dpcategorieslist}{fielddescriptor}(x, datapackage = dpgetdatapackage(x), normalised = FALSE, ...)
}
\arguments{
\item{x}{the variable for which to get the Categories List}

\item{...}{used to pass extra arguments on to other methods.}

\item{fielddescriptor}{the Field Descriptor associated with the variable.}

\item{datapackage}{the Data Package where the variable is from.}

\item{normalised}{if \code{TRUE} the column with values will be named 
\code{value} and the the columnd with labels \code{label}.}
}
\value{
Returns a \code{data.frame} with the categories or \code{NULL} when none could
be found.
}
\description{
Get the a data.frame with the categories for a variable.
}
