% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcheckdataresource.R, R/dpcheckfield.R
\name{dpcheckdataresource}
\alias{dpcheckdataresource}
\alias{dpcheckfield}
\title{Check is dataset is valid given Data Resource}
\usage{
dpcheckdataresource(
  x,
  dataresource = attr(x, "resource"),
  constraints = TRUE,
  throw = FALSE,
  tolerance = sqrt(.Machine$double.eps)
)

dpcheckfield(
  x,
  fielddescriptor,
  constraints = TRUE,
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{vector to test}

\item{dataresource}{\code{dataresource} object to check \code{x} against.}

\item{constraints}{also check relevant constraints in the field descriptor.}

\item{throw}{generate an error if the data set is not valid according to the 
dataresource.}

\item{tolerance}{numerical tolerance used in some of the tests}

\item{fielddescriptor}{field descriptor to test the vector against}
}
\value{
Returns \code{TRUE} when the field is valid. Returns a character vector with
length >= 1 if the field is not valid. The text in the character values 
indicates why the field is not valid.

When \code{throw = TRUE} the function will generate an error instead of
returning a character vector. When the dataset is valid the function returns
\code{TRUE} invisibly.

Returns \code{TRUE} when the field is valid. Returns a character vector with
length >= 1 if the field is not valid. The text in the character values 
indicates why the field is not valid.
}
\description{
Check is dataset is valid given Data Resource

Check if a vector is valid given a field descriptor
}
\seealso{
Use \code{\link{isTRUE}} to check if the test was successful.

Use \code{\link{isTRUE}} to check if the test was successful.
}
