% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpfield.R
\name{dpfield}
\alias{dpfield}
\title{Get the Field Descriptor associated with a certain field in a Data Resource}
\usage{
dpfield(x, fieldname)
}
\arguments{
\item{x}{a \code{dataresource} or \code{tableschema} object.}

\item{fieldname}{length one character vector with the name of the field.}
}
\value{
An object of type \code{fielddescriptor}.
}
\description{
Get the Field Descriptor associated with a certain field in a Data Resource
}
