% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpfieldnames.R
\name{dpfieldnames}
\alias{dpfieldnames}
\title{List the fields in a Data Resource}
\usage{
dpfieldnames(x)
}
\arguments{
\item{x}{object for which to get the field names. This can either be a Data
Resource or Table Schema.}
}
\value{
Returns a character vector with the fields in the Data Resource.
}
\description{
List the fields in a Data Resource
}
