% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgeneratedataresources.R
\name{dpgeneratedataresources}
\alias{dpgeneratedataresources}
\title{Generate Data Resources for a dataset}
\usage{
dpgeneratedataresources(
  x,
  name,
  path = paste0(name, getextension(format)),
  format = "csv",
  mediatype = getmediatype(format),
  categories_type = c("regular", "resource"),
  format_codelists = format,
  mediatype_codelists = getmediatype(format_codelists),
  ...
)
}
\arguments{
\item{x}{\code{data.frame} for which to generate the Data Resources.}

\item{name}{name of the Data Resource}

\item{path}{name of the file in which to store the dataset. This should be a
path relative to the location of the directory in which the Data Package in
which the Data Resource will be stored.}

\item{format}{the data format in which the data is stored.}

\item{mediatype}{mediatype of the data}

\item{categories_type}{how should categories be stored. See 
\code{\link{dpgeneratefielddescriptor}}.}

\item{format_codelists}{data format to use for the code lists}

\item{mediatype_codelists}{mediatyp of the code lists}

\item{...}{Currently ignored}
}
\value{
Returns a \code{list} with Data Resource objects. The first Data Resource is
that of the dataset \code{x} itself. Additional Data Resources are for the
code lists of the variables. Code lists are generated for fields that have a
code list and for \code{factor} variables.

Note that this function does not create the file at \code{path}. The export
of the Data Resource is automatically set to CSV.
}
\description{
Generate Data Resources for a dataset
}
\examples{
# generate an example dataset
dta <- data.frame(a = 1:3, b = factor(letters[1:3]))
resources <- dpgeneratedataresources(dta, "example")
print(resources)

}
