% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgetconnection.R
\name{dpgetconnection}
\alias{dpgetconnection}
\title{Get a connection to the data belonging to a Data Resource}
\usage{
dpgetconnection(x, ...)
}
\arguments{
\item{x}{Can either be a Data Resource or Data Package.}

\item{...}{Extra arguments are passed on to \code{\link{dpgetdata}}.}
}
\value{
Depending on the type of Data Resource a connection to the data is returned.
Some readers will return the data set as a \code{data.frame}.
}
\description{
Get a connection to the data belonging to a Data Resource
}
\details{
When \code{x} is a Data Package a additional argument \code{resourcename} is
needed to identify the correct Data Resource. See \code{\link{dpgetdata}}. 

This function calls \code{dpgetdata} with an additional 
\code{as_connection = TRUE)} argument.
}
