% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpnresources.R
\name{dpnresources}
\alias{dpnresources}
\title{Return the number of resources in a Data Package}
\usage{
dpnresources(dp)
}
\arguments{
\item{dp}{A Data Package object.}
}
\value{
Returns an integer with the number of resources in the Data Package.
}
\description{
Return the number of resources in a Data Package
}
