% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpproperties.R
\name{dpproperties}
\alias{dpproperties}
\alias{dpproperties.readonlydatapackage}
\alias{dpproperties.editabledatapackage}
\alias{dpproperties.dataresource}
\alias{dpproperties.tableschema}
\title{Get a list of properties defined for the object}
\usage{
dpproperties(x)

\method{dpproperties}{readonlydatapackage}(x)

\method{dpproperties}{editabledatapackage}(x)

\method{dpproperties}{dataresource}(x)

\method{dpproperties}{tableschema}(x)
}
\arguments{
\item{x}{the object for which to obtain the properties}
}
\value{
Returns a character vector (possibly zero length) with the names of the
properties.
}
\description{
Get a list of properties defined for the object
}
\seealso{
The \code{\link{dpproperty}} method can be used to get the values of the
properties.
}
