% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpproperty.R
\name{dpproperty}
\alias{dpproperty}
\alias{dpproperty.readonlydatapackage}
\alias{dpproperty.editabledatapackage}
\alias{dpproperty<-}
\alias{dpproperty<-.readonlydatapackage}
\alias{dpproperty<-.editabledatapackage}
\alias{dpproperty.dataresource}
\alias{dpproperty<-.dataresource}
\alias{dpproperty.tableschema}
\alias{dpproperty<-.tableschema}
\alias{dpproperty.fielddescriptor}
\alias{dpproperty<-.fielddescriptor}
\title{Get and set properties of Data Packages and Data Resources}
\usage{
dpproperty(x, attribute)

\method{dpproperty}{readonlydatapackage}(x, attribute)

\method{dpproperty}{editabledatapackage}(x, attribute)

dpproperty(x, attribute) <- value

\method{dpproperty}{readonlydatapackage}(x, attribute) <- value

\method{dpproperty}{editabledatapackage}(x, attribute) <- value

\method{dpproperty}{dataresource}(x, attribute)

\method{dpproperty}{dataresource}(x, attribute) <- value

\method{dpproperty}{tableschema}(x, attribute)

\method{dpproperty}{tableschema}(x, attribute) <- value

\method{dpproperty}{fielddescriptor}(x, attribute)

\method{dpproperty}{fielddescriptor}(x, attribute) <- value
}
\arguments{
\item{x}{a \code{datapackage} or \code{dataresource} object.}

\item{attribute}{a length 1 character vector with the name of the property.}

\item{value}{the new value of the property.}
}
\value{
Either returns the property or modifies the object.
}
\description{
Get and set properties of Data Packages and Data Resources
}
\seealso{
See \code{\link{dpname}} etc. for methods for specific properties for Data
Packages and \code{\link{dpencoding}} etc. for specific properties for Data
Resources. These specific methods also check if the input is valid for the
given property.
}
