% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpresource.R
\name{dpresource}
\alias{dpresource}
\alias{dpresource.datapackage}
\alias{dpresource<-}
\alias{dpresource<-.readonlydatapackage}
\alias{dpresource<-.editabledatapackage}
\title{Modifying the resources of a Data Package}
\usage{
dpresource(x, resourcename)

\method{dpresource}{datapackage}(x, resourcename)

dpresource(x, resourcename) <- value

\method{dpresource}{readonlydatapackage}(x, resourcename) <- value

\method{dpresource}{editabledatapackage}(x, resourcename) <- value
}
\arguments{
\item{x}{a \code{datapackage} object.}

\item{resourcename}{the name of a resource.}

\item{value}{a \code{dataresource} object.}
}
\value{
Either returns a Data Resource object or modifies the Data Package.
}
\description{
Modifying the resources of a Data Package
}
\details{
When a resource with the name already exists this resource is
overwritten. Therefore, the assignment operator can also be used to modify
existing resources.
}
