% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpresourcenames.R
\name{dpresourcenames}
\alias{dpresourcenames}
\title{Get the names of the resources in a Data Package}
\usage{
dpresourcenames(dp)
}
\arguments{
\item{dp}{A \code{datapackage} object.}
}
\value{
Returns a character vector with the names of the data resources in the Data
Package.
}
\description{
Get the names of the resources in a Data Package
}
