% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpresources.R
\name{dpresources<-}
\alias{dpresources<-}
\title{Modify a set of Data Resources in a Data Package}
\usage{
dpresources(x) <- value
}
\arguments{
\item{x}{a \code{datapackage} object}

\item{value}{a list of \code{dataresource} objects}
}
\value{
Returns a modified \code{x}.
}
\description{
Modify a set of Data Resources in a Data Package
}
