% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpsaveasdatapackage.R
\name{dpsaveasdatapackage}
\alias{dpsaveasdatapackage}
\title{Save a dataset as a Data Package}
\usage{
dpsaveasdatapackage(
  data,
  path,
  name,
  categories_type = c("regular", "resource")
)
}
\arguments{
\item{data}{the data.frame with the data to save}

\item{path}{directory in which to create the datapackage}

\item{name}{name of the Data Resource. When omitted a name is generated.}

\item{categories_type}{how should categories be stored. See 
\code{\link{dpgeneratefielddescriptor}}.}
}
\value{
Does not return anything. Called for the side effect of creating a directory
and creating a number of files in this directory. Together these form a
complete Data Package.
}
\description{
Save a dataset as a Data Package
}
\details{
This function is a wrapper function around \code{\link{newdatapackage}},
\code{\link{dpgeneratedataresource}} and \code{\link{dpwritedata}}. These
functions are called with the default arguments. This allows for a quick way
to save a data set with any necessary data needed to read the dataset.
}
