% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema}
\alias{schema}
\alias{schema.default}
\alias{schema<-}
\alias{schema<-.default}
\title{Get and set the schema of a data set or columns in the data set}
\usage{
schema(x, attribute, ...)

\method{schema}{default}(x, attribute, ...)

schema(x, attribute) <- value

\method{schema}{default}(x, attribute) <- value
}
\arguments{
\item{x}{the data.frame or vector for which the schema should be set or
obtained.}

\item{attribute}{the optional name of the specific attribute to get or set.
When omitted the complete schema is set or obtained.}

\item{...}{passed on to other methods.}

\item{value}{the value to set the schema or attribute with.}
}
\description{
Get and set the schema of a data set or columns in the data set
}
\details{
The schema is stored in the \code{schema} attribute of the object. When the
object does not have a schema associated with it, a default schema appropiate
for the object is generated.
}
