% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwf_reader.R
\name{fwf_reader}
\alias{fwf_reader}
\title{Read the FWF-data for a Data Resource}
\usage{
fwf_reader(path, resource, to_factor = FALSE, ...)
}
\arguments{
\item{path}{path to the data set.}

\item{resource}{a Data Resource.}

\item{to_factor}{convert columns to factor if the schema has a categories
field for the column. Passed on to \code{\link{dpapplyschema}}.}

\item{...}{additional arguments are passed on to \code{\link{dpapplyschema}}.}
}
\value{
Returns a \code{data.frame} with the data.
}
\description{
Read the FWF-data for a Data Resource
}
\seealso{
Generally used by calling \code{\link{dpgetdata}}.
}
