% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R, R/properties_datapackage.R
\name{dpcontributors}
\alias{dpcontributors}
\alias{dpcontributors<-}
\alias{dpcontributors.datapackage}
\alias{dpcontributors<-.datapackage}
\alias{dpname}
\alias{dpname.datapackage}
\alias{dpname<-}
\alias{dpname<-.datapackage}
\alias{dptitle}
\alias{dptitle.datapackage}
\alias{dptitle<-}
\alias{dptitle<-.datapackage}
\alias{dpdescription}
\alias{dpdescription.datapackage}
\alias{dpdescription<-}
\alias{dpdescription<-.datapackage}
\alias{dpkeywords}
\alias{dpkeywords.datapackage}
\alias{dpkeywords<-}
\alias{dpkeywords<-.datapackage}
\alias{dpcreated}
\alias{dpcreated.datapackage}
\alias{dpcreated<-}
\alias{dpcreated<-.datapackage}
\alias{dpid}
\alias{dpid.datapackage}
\alias{dpid<-}
\alias{dpid<-.datapackage}
\title{Getting and setting properties of Data Packages}
\usage{
dpcontributors(x, ...)

dpcontributors(x) <- value

\method{dpcontributors}{datapackage}(x, ...)

\method{dpcontributors}{datapackage}(x) <- value

dpname(x)

\method{dpname}{datapackage}(x)

dpname(x) <- value

\method{dpname}{datapackage}(x) <- value

dptitle(x)

\method{dptitle}{datapackage}(x)

dptitle(x) <- value

\method{dptitle}{datapackage}(x) <- value

dpdescription(x, ..., firstparagraph = FALSE, dots = FALSE)

\method{dpdescription}{datapackage}(x, ..., firstparagraph = FALSE, dots = FALSE)

dpdescription(x) <- value

\method{dpdescription}{datapackage}(x) <- value

dpkeywords(x, ...)

\method{dpkeywords}{datapackage}(x, ...)

dpkeywords(x) <- value

\method{dpkeywords}{datapackage}(x) <- value

dpcreated(x, ...)

\method{dpcreated}{datapackage}(x, ...)

dpcreated(x) <- value

\method{dpcreated}{datapackage}(x) <- value

dpid(x, ...)

\method{dpid}{datapackage}(x, ...)

dpid(x) <- value

\method{dpid}{datapackage}(x) <- value
}
\arguments{
\item{x}{a \code{datapackage} object.}

\item{...}{used to pass additional arguments to other methods.}

\item{value}{the new value of the property.}

\item{firstparagraph}{Only return the first paragraph of the description.}

\item{dots}{When returning only the first paragraph indicate missing
paragraphs with \code{...}.}
}
\value{
Either returns the property or modifies the object.
}
\description{
Getting and setting properties of Data Packages
}
\seealso{
See \code{\link{dp_resource}} for methods for getting and setting the resources
of a Data Package.
}
