% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties_dataresource.R,
%   R/properties_tabulardataresource.R
\name{dpname.dataresource}
\alias{dpname.dataresource}
\alias{dpname<-.dataresource}
\alias{dptitle.dataresource}
\alias{dptitle<-.dataresource}
\alias{dpdescription.dataresource}
\alias{dpdescription<-.dataresource}
\alias{dppath}
\alias{dppath<-}
\alias{dppath.dataresource}
\alias{dppath<-.dataresource}
\alias{dpformat}
\alias{dpformat<-}
\alias{dpformat.dataresource}
\alias{dpformat<-.dataresource}
\alias{dpmediatype}
\alias{dpmediatype<-}
\alias{dpmediatype.dataresource}
\alias{dpmediatype<-.dataresource}
\alias{dpencoding}
\alias{dpencoding<-}
\alias{dpencoding.dataresource}
\alias{dpencoding<-.dataresource}
\alias{dpbytes}
\alias{dpbytes<-}
\alias{dpbytes.dataresource}
\alias{dpbytes<-.dataresource}
\alias{dphash}
\alias{dphash<-}
\alias{dphash.dataresource}
\alias{dphash<-.dataresource}
\alias{dpschema}
\alias{dpschema.dataresource}
\title{Getting and setting properties of Data Resources}
\usage{
\method{dpname}{dataresource}(x)

\method{dpname}{dataresource}(x) <- value

\method{dptitle}{dataresource}(x)

\method{dptitle}{dataresource}(x) <- value

\method{dpdescription}{dataresource}(x, ..., firstparagraph = FALSE, dots = FALSE)

\method{dpdescription}{dataresource}(x) <- value

dppath(x, ...)

dppath(x) <- value

\method{dppath}{dataresource}(x, fullpath = FALSE, ...)

\method{dppath}{dataresource}(x) <- value

dpformat(x, ...)

dpformat(x) <- value

\method{dpformat}{dataresource}(x, default = FALSE, ...)

\method{dpformat}{dataresource}(x) <- value

dpmediatype(x, ...)

dpmediatype(x) <- value

\method{dpmediatype}{dataresource}(x, ...)

\method{dpmediatype}{dataresource}(x) <- value

dpencoding(x, default = FALSE, ...)

dpencoding(x) <- value

\method{dpencoding}{dataresource}(x, default = FALSE, ...)

\method{dpencoding}{dataresource}(x) <- value

dpbytes(x, ...)

dpbytes(x) <- value

\method{dpbytes}{dataresource}(x, ...)

\method{dpbytes}{dataresource}(x) <- value

dphash(x, ...)

dphash(x) <- value

\method{dphash}{dataresource}(x, ...)

\method{dphash}{dataresource}(x) <- value

dpschema(x)

\method{dpschema}{dataresource}(x)
}
\arguments{
\item{x}{a \code{dataresource} object.}

\item{value}{the new value of the property.}

\item{...}{used to pass additional arguments to other methods.}

\item{firstparagraph}{Only return the first paragraph of the description.}

\item{dots}{When returning only the first paragraph indicate missing
paragraphs with \code{...}.}

\item{fullpath}{Return the full path including the path to the Data Package
and not only the path relative to the Data Package. This is only relevant for
relative paths.}

\item{default}{return the default value if the property had a default value 
and the property is not set.}
}
\value{
Either returns the property or modifies the object. If the property of not
set \code{NULL} is returned (unless \code{default = TRUE}).
}
\description{
Getting and setting properties of Data Resources
}
