% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties_fielddescriptor.R
\name{PropertiesFielddescriptor}
\alias{PropertiesFielddescriptor}
\alias{dp_name.fielddescriptor}
\alias{dp_title.fielddescriptor}
\alias{dp_description<-.fielddescriptor}
\alias{dp_format.fielddescriptor}
\alias{dp_type}
\alias{dp_type<-}
\alias{dp_type.fielddescriptor}
\title{Getting and setting properties of Data Resources}
\usage{
\method{dp_name}{fielddescriptor}(x)

\method{dp_title}{fielddescriptor}(x)

\method{dp_description}{fielddescriptor}(x) <- value

\method{dp_format}{fielddescriptor}(x, ...)

dp_type(x)

dp_type(x) <- value

\method{dp_type}{fielddescriptor}(x)
}
\arguments{
\item{x}{a \code{fielddescriptor} object.}

\item{value}{the new value of the property.}

\item{...}{used to pass additional arguments to other methods.}
}
\value{
Either returns the property or modifies the object. If the property is not
set \code{NULL} is returned (unless \code{default = TRUE}).
}
\description{
Getting and setting properties of Data Resources
}
\seealso{
See \link{PropertiesDatapackage} and \link{PropertiesDataresource} for methods
for Data Packages and Data Resources respectively. Also see
\code{\link{dp_property}} for a generic method for getting and setting
properties. These functions can also be used to get and set 'unofficial'
properties'
}
