% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledatapackage-package.R
\docType{package}
\name{sampledatapackage-package}
\alias{sampledatapackage}
\alias{sampledatapackage-package}
\title{sampledatapackage: Data package with shrimp prices}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
