% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_boolean.R
\name{to_boolean}
\alias{to_boolean}
\title{Convert a vector to 'boolean' using the specified schema}
\usage{
to_boolean(x, schema = list(), ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{schema}{the table-schema for the field.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{logical} vector with \code{schema} added as the 'schema'
attribute.
}
\description{
Convert a vector to 'boolean' using the specified schema
}
\details{
When \code{schema} is missing a default schema is generated using
\code{\link{complete_schema_boolean}}.
}
