% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_date.R
\name{to_date}
\alias{to_date}
\title{Convert a vector to 'date' using the specified schema}
\usage{
to_date(x, schema = list(), ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{schema}{the table-schema for the field.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{Date} vector with \code{schema} added as the 'schema'
attribute.
}
\description{
Convert a vector to 'date' using the specified schema
}
\details{
When \code{schema} is missing a default schema is generated using
\code{\link{complete_schema_date}}.
}
