% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_integer.R
\name{to_integer}
\alias{to_integer}
\title{Convert a vector to 'integer' using the specified schema}
\usage{
to_integer(x, schema = list(), ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{schema}{the table-schema for the field.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{integer} vector with \code{schema} added as the 'schema'
attribute.
}
\description{
Convert a vector to 'integer' using the specified schema
}
\details{
When \code{schema} is missing a default schema is generated using
\code{\link{complete_schema_integer}}.
}
