% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_number.R
\name{to_number}
\alias{to_number}
\title{Convert a vector to 'number' using the specified schema}
\usage{
to_number(x, schema = list(), decimalChar = ".", ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{schema}{the field schema for the field.}

\item{decimalChar}{decimal separator. Used when the field schema does not
specify a decimal separator.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{numeric} vector with \code{schema} added as the 'schema'
attribute.
}
\description{
Convert a vector to 'number' using the specified schema
}
\details{
When \code{schema} is missing a default schema is generated using
\code{\link{complete_schema_number}}.
}
