% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcodelist.R
\name{dpcodelist}
\alias{dpcodelist}
\alias{dpcodelist.default}
\alias{dpcodelist.fielddescriptor}
\title{Get the Code List for a variable.}
\usage{
dpcodelist(x, ...)

\method{dpcodelist}{default}(
  x,
  fielddescriptor = attr(x, "fielddescriptor"),
  datapackage = dpgetdatapackage(fielddescriptor),
  ...
)

\method{dpcodelist}{fielddescriptor}(x, datapackage = dpgetdatapackage(x), normalised = FALSE, ...)
}
\arguments{
\item{x}{the variable for which to get the Code List}

\item{...}{used to pass extra arguments on to other methods.}

\item{fielddescriptor}{the Field Descriptor associated with the variable.}

\item{datapackage}{the Data Package where the variable is from.}

\item{normalised}{if \code{TRUE} the column with values will be named 
\code{value} and the the columnd with labels \code{label}.}
}
\value{
Returns a \code{data.frame} with the Code List or \code{NULL} when none could
be found.
}
\description{
Get the Code List for a variable.
}
