% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgeneratefielddescriptor.R
\name{dpgeneratefielddescriptor}
\alias{dpgeneratefielddescriptor}
\alias{dpgeneratefielddescriptor.default}
\alias{dpgeneratefielddescriptor.numeric}
\alias{dpgeneratefielddescriptor.integer}
\alias{dpgeneratefielddescriptor.logical}
\alias{dpgeneratefielddescriptor.Date}
\alias{dpgeneratefielddescriptor.character}
\alias{dpgeneratefielddescriptor.factor}
\title{Generate a fielddescriptor for a given variable in a dataset}
\usage{
dpgeneratefielddescriptor(x, name, ...)

\method{dpgeneratefielddescriptor}{default}(x, name, ...)

\method{dpgeneratefielddescriptor}{numeric}(
  x,
  name,
  use_existing = TRUE,
  use_categories = TRUE,
  categories_type = c("regular", "resource"),
  ...
)

\method{dpgeneratefielddescriptor}{integer}(
  x,
  name,
  use_existing = TRUE,
  use_categories = TRUE,
  categories_type = c("regular", "resource"),
  ...
)

\method{dpgeneratefielddescriptor}{logical}(
  x,
  name,
  use_existing = TRUE,
  use_categories = TRUE,
  categories_type = c("regular", "resource"),
  ...
)

\method{dpgeneratefielddescriptor}{Date}(
  x,
  name,
  use_existing = TRUE,
  use_categories = TRUE,
  categories_type = c("regular", "resource"),
  ...
)

\method{dpgeneratefielddescriptor}{character}(
  x,
  name,
  use_existing = TRUE,
  use_categories = TRUE,
  categories_type = c("regular", "resource"),
  ...
)

\method{dpgeneratefielddescriptor}{factor}(
  x,
  name,
  use_existing = TRUE,
  use_categories = TRUE,
  categories_type = c("regular", "resource"),
  ...
)
}
\arguments{
\item{x}{vector for which to generate the fielddescriptor}

\item{name}{name of the field in the dataset.}

\item{...}{used to pass extra arguments to methods.}

\item{use_existing}{use existing field descriptor if present (assumes this is
stored in the 'fielddescriptor' attribute).}

\item{use_categories}{do not generate a categories field except when \code{x}
is a factor.}

\item{categories_type}{how should categories be stored. Note that type "resource"
is not officially part of the standard.}
}
\value{
Returns a \code{fielddescriptor}.
}
\description{
Generate a fielddescriptor for a given variable in a dataset
}
