% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dptofactor.R
\name{dptofactor}
\alias{dptofactor}
\title{Recode a variable to factor using the associated Code List}
\usage{
dptofactor(x, codelist = dpcodelist(x), warn = TRUE)
}
\arguments{
\item{x}{the variable to recode}

\item{codelist}{a \code{data.frame} with the Code List. It is assumed that
the first column contains the codes and the second column the labels.}

\item{warn}{give a warning when there is no code list.}
}
\value{
Returns a factor vector or \code{x} when no Code List could be found
(\code{codelist = NULL}).
}
\description{
Recode a variable to factor using the associated Code List
}
