% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwf_reader.R
\name{fwf_reader}
\alias{fwf_reader}
\title{Read the FWF-data for a Data Resource}
\usage{
fwf_reader(path, resource, convert_categories = c("no", "to_factor"), ...)
}
\arguments{
\item{path}{path to the data set.}

\item{resource}{a Data Resource.}

\item{convert_categories}{how to handle columns for which the field
descriptor has a \code{categories} property. Passed on to
\code{\link{dp_apply_schema}}.}

\item{...}{additional arguments are passed on to \code{\link{dp_apply_schema}}.}
}
\value{
Returns a \code{data.frame} with the data.
}
\description{
Read the FWF-data for a Data Resource
}
\seealso{
Generally used by calling \code{\link{dp_get_data}}.
}
