% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdataresource.R
\name{newdataresource}
\alias{newdataresource}
\title{Create a new Data Resource}
\usage{
newdataresource(name, title = NULL, description = NULL, ...)
}
\arguments{
\item{name}{The name of the Data Resource.}

\item{title}{The title of the Data Resource.}

\item{description}{The description of the Data Resource.}

\item{...}{Ignored for now.}
}
\value{
Returns a \code{dataresource} object which is a list with the properties of
the Data Resource.
}
\description{
Create a new Data Resource
}
\examples{
dir <- tempdir()
dp <- newdatapackage(dir, name = "test-package")

res <- newdataresource(name = "iris")
dptitle(res) <- "The Iris Data Set"
dpencoding(res) <- "UTF-8"
dpmediatype(res) <- "text/csv"

# resource adds a resource if it doesn't yet exist or updates
# an existing resource
dpresource(dp, "iris") <- res

\dontshow{
file.remove(file.path(dir, "datapackage.json"))
file.remove(dir)
}


}
