% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R, R/properties_datapackage.R
\name{dp_contributors}
\alias{dp_contributors}
\alias{dp_contributors<-}
\alias{dp_contributors.datapackage}
\alias{dp_contributors<-.datapackage}
\alias{dp_name}
\alias{dp_name.datapackage}
\alias{dp_name<-}
\alias{dp_name<-.datapackage}
\alias{dp_title}
\alias{dp_title.datapackage}
\alias{dp_title<-}
\alias{dp_title<-.datapackage}
\alias{dp_description}
\alias{dp_description.datapackage}
\alias{dp_description<-}
\alias{dp_description<-.datapackage}
\alias{dp_keywords}
\alias{dp_keywords.datapackage}
\alias{dp_keywords<-}
\alias{dp_keywords<-.datapackage}
\alias{dp_created}
\alias{dp_created.datapackage}
\alias{dp_created<-}
\alias{dp_created<-.datapackage}
\alias{dp_id}
\alias{dp_id.datapackage}
\alias{dp_id<-}
\alias{dp_id<-.datapackage}
\title{Getting and setting properties of Data Packages}
\usage{
dp_contributors(x, ...)

dp_contributors(x) <- value

\method{dp_contributors}{datapackage}(x, ...)

\method{dp_contributors}{datapackage}(x) <- value

dp_name(x)

\method{dp_name}{datapackage}(x)

dp_name(x) <- value

\method{dp_name}{datapackage}(x) <- value

dp_title(x)

\method{dp_title}{datapackage}(x)

dp_title(x) <- value

\method{dp_title}{datapackage}(x) <- value

dp_description(x, ..., firstparagraph = FALSE, dots = FALSE)

\method{dp_description}{datapackage}(x, ..., firstparagraph = FALSE, dots = FALSE)

dp_description(x) <- value

\method{dp_description}{datapackage}(x) <- value

dp_keywords(x, ...)

\method{dp_keywords}{datapackage}(x, ...)

dp_keywords(x) <- value

\method{dp_keywords}{datapackage}(x) <- value

dp_created(x, ...)

\method{dp_created}{datapackage}(x, ...)

dp_created(x) <- value

\method{dp_created}{datapackage}(x) <- value

dp_id(x, ...)

\method{dp_id}{datapackage}(x, ...)

dp_id(x) <- value

\method{dp_id}{datapackage}(x) <- value
}
\arguments{
\item{x}{a \code{datapackage} object.}

\item{...}{used to pass additional arguments to other methods.}

\item{value}{the new value of the property.}

\item{firstparagraph}{Only return the first paragraph of the description.}

\item{dots}{When returning only the first paragraph indicate missing
paragraphs with \code{...}.}
}
\value{
Either returns the property or modifies the object.
}
\description{
Getting and setting properties of Data Packages
}
\seealso{
See \code{\link{dp_resource}} for methods for getting and setting the resources
of a Data Package.
}
