% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties_dataresource.R,
%   R/properties_fielddescriptor.R, R/properties_tabulardataresource.R
\name{dp_name.dataresource}
\alias{dp_name.dataresource}
\alias{dp_name<-.dataresource}
\alias{dp_title.dataresource}
\alias{dp_title<-.dataresource}
\alias{dp_description.dataresource}
\alias{dp_description<-.dataresource}
\alias{dp_path}
\alias{dp_path<-}
\alias{dp_path.dataresource}
\alias{dp_path<-.dataresource}
\alias{dp_format}
\alias{dp_format<-}
\alias{dp_format.dataresource}
\alias{dp_format<-.dataresource}
\alias{dp_mediatype}
\alias{dp_mediatype<-}
\alias{dp_mediatype.dataresource}
\alias{dp_mediatype<-.dataresource}
\alias{dp_encoding}
\alias{dp_encoding<-}
\alias{dp_encoding.dataresource}
\alias{dp_encoding<-.dataresource}
\alias{dp_bytes}
\alias{dp_bytes<-}
\alias{dp_bytes.dataresource}
\alias{dp_bytes<-.dataresource}
\alias{dp_hash}
\alias{dp_hash<-}
\alias{dp_hash.dataresource}
\alias{dp_hash<-.dataresource}
\alias{dp_name<-.fielddescriptor}
\alias{dp_title<-.fielddescriptor}
\alias{dp_description.fielddescriptor}
\alias{dp_format<-.fielddescriptor}
\alias{dp_schema}
\alias{dp_schema.dataresource}
\title{Getting and setting properties of Data Resources}
\usage{
\method{dp_name}{dataresource}(x)

\method{dp_name}{dataresource}(x) <- value

\method{dp_title}{dataresource}(x)

\method{dp_title}{dataresource}(x) <- value

\method{dp_description}{dataresource}(x, ..., first_paragraph = FALSE, dots = FALSE)

\method{dp_description}{dataresource}(x) <- value

dp_path(x, ...)

dp_path(x) <- value

\method{dp_path}{dataresource}(x, full_path = FALSE, ...)

\method{dp_path}{dataresource}(x) <- value

dp_format(x, ...)

dp_format(x) <- value

\method{dp_format}{dataresource}(x, default = FALSE, ...)

\method{dp_format}{dataresource}(x) <- value

dp_mediatype(x, ...)

dp_mediatype(x) <- value

\method{dp_mediatype}{dataresource}(x, ...)

\method{dp_mediatype}{dataresource}(x) <- value

dp_encoding(x, default = FALSE, ...)

dp_encoding(x) <- value

\method{dp_encoding}{dataresource}(x, default = FALSE, ...)

\method{dp_encoding}{dataresource}(x) <- value

dp_bytes(x, ...)

dp_bytes(x) <- value

\method{dp_bytes}{dataresource}(x, ...)

\method{dp_bytes}{dataresource}(x) <- value

dp_hash(x, ...)

dp_hash(x) <- value

\method{dp_hash}{dataresource}(x, ...)

\method{dp_hash}{dataresource}(x) <- value

\method{dp_name}{fielddescriptor}(x) <- value

\method{dp_title}{fielddescriptor}(x) <- value

\method{dp_description}{fielddescriptor}(x, ..., first_paragraph = FALSE, dots = FALSE)

\method{dp_format}{fielddescriptor}(x) <- value

dp_schema(x)

\method{dp_schema}{dataresource}(x)
}
\arguments{
\item{x}{a \code{dataresource} object.}

\item{value}{the new value of the property.}

\item{...}{used to pass additional arguments to other methods.}

\item{first_paragraph}{Only return the first paragraph of the description.}

\item{dots}{When returning only the first paragraph indicate missing
paragraphs with \code{...}.}

\item{full_path}{Return the full path including the path to the Data Package
and not only the path relative to the Data Package. This is only relevant for
relative paths.}

\item{default}{return the default value if the property had a default value 
and the property is not set.}
}
\value{
Either returns the property or modifies the object. If the property of not
set \code{NULL} is returned (unless \code{default = TRUE}).
}
\description{
Getting and setting properties of Data Resources
}
