% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties_fielddescriptor.R
\name{dp_name.fielddescriptor}
\alias{dp_name.fielddescriptor}
\alias{dp_title.fielddescriptor}
\alias{dp_description<-.fielddescriptor}
\alias{dp_format.fielddescriptor}
\title{Getting and setting properties of Data Resources}
\usage{
\method{dp_name}{fielddescriptor}(x)

\method{dp_title}{fielddescriptor}(x)

\method{dp_description}{fielddescriptor}(x) <- value

\method{dp_format}{fielddescriptor}(x, ...)
}
\arguments{
\item{x}{a \code{fielddescriptor} object.}

\item{value}{the new value of the property.}

\item{...}{used to pass additional arguments to other methods.}
}
\value{
Either returns the property or modifies the object. If the property is not
set \code{NULL} is returned (unless \code{default = TRUE}).
}
\description{
Getting and setting properties of Data Resources
}
