% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_integer.R
\name{to_integer}
\alias{to_integer}
\title{Convert a vector to 'integer' using the specified field descriptor}
\usage{
to_integer(x, fielddescriptor = list(), ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{fielddescriptor}{the field descriptor for the field.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{integer} vector with \code{fielddescriptor} added as
the 'fielddescriptor' attribute.
}
\description{
Convert a vector to 'integer' using the specified field descriptor
}
\details{
When \code{fielddescriptor} is missing a default field descriptor is
generated using \code{\link{complete_fielddescriptor_integer}}.
}
