% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_number.R
\name{to_number}
\alias{to_number}
\title{Convert a vector to 'number' using the specified field descriptor}
\usage{
to_number(x, fielddescriptor = list(), decimalChar = ".", ...)
}
\arguments{
\item{x}{the vector to convert.}

\item{fielddescriptor}{the field descriptor for the field.}

\item{decimalChar}{decimal separator. Used when the field field descriptor
does not specify a decimal separator.}

\item{...}{passed on to other methods.}
}
\value{
Will return an \code{numeric} vector with \code{fielddescriptor} added as
the 'fielddescriptor' attribute.
}
\description{
Convert a vector to 'number' using the specified field descriptor
}
\details{
When \code{fielddescriptor} is missing a default field descriptor is
generated using \code{\link{complete_fielddescriptor_number}}.
}
